requirements.txt
py-solc
py-etherscan-api
requests